Request, Response -> DTO 통합
