+++ src/main/kotlin/mara/server/domain/ingredient/IngredientDto.kt
@@ -1,5 +1,12 @@
 package mara.server.domain.ingredient
 
+data class IngredientRequest(
+    val category: String,
+    val name: String,
+    val iconImage: String,
+    val expirationDays: Int
+)
+
 data class IngredientResponse(
     val ingredientId: Long,
     val category: String,
