--- src/main/kotlin/mara/server/domain/ingredient/IngredientRequest.kt
@@ -1,8 +1,0 @@
-package mara.server.domain.ingredient
-
-data class IngredientRequest(
-    val category: String,
-    val name: String,
-    val iconImage: String,
-    val expirationDays: Int
-)
